Renamed aggregate_functions.rb to functions.rb, to preserve the file_name -> NameSpace mapping.
